requirements.txt
imantics
matplotlib
opencv-python
numpy
Pillow
tqdm
pytest
